# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students": {
      "get": {
        "operationId": "StudentsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search keyword applied to nationalId, firstName, and lastName",
            "schema": {
              "example": "Juana",
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "Filters students enrolled in the provided school year ID",
            "schema": {
              "minimum": 1,
              "example": 2025,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "post": {
        "operationId": "StudentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/students/{id}": {
      "get": {
        "operationId": "StudentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "patch": {
        "operationId": "StudentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      },
      "delete": {
        "operationId": "StudentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "students"
        ]
      }
    },
    "/subject-areas": {
      "get": {
        "operationId": "SubjectAreasController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search keyword applied to area code and name",
            "schema": {
              "example": "math",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subject-areas"
        ]
      },
      "post": {
        "operationId": "SubjectAreasController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectAreaDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subject-areas"
        ]
      }
    },
    "/subject-areas/{id}": {
      "get": {
        "operationId": "SubjectAreasController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subject-areas"
        ]
      },
      "patch": {
        "operationId": "SubjectAreasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectAreaDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subject-areas"
        ]
      },
      "delete": {
        "operationId": "SubjectAreasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subject-areas"
        ]
      }
    },
    "/subjects": {
      "get": {
        "operationId": "SubjectsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search keyword applied to subject code and name",
            "schema": {
              "example": "algebra",
              "type": "string"
            }
          },
          {
            "name": "areaId",
            "required": false,
            "in": "query",
            "description": "Restrict results to a specific subject area",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subjects"
        ]
      },
      "post": {
        "operationId": "SubjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubjectDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subjects"
        ]
      }
    },
    "/subjects/{id}": {
      "get": {
        "operationId": "SubjectsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subjects"
        ]
      },
      "patch": {
        "operationId": "SubjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubjectDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subjects"
        ]
      },
      "delete": {
        "operationId": "SubjectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden: requires role admin, coordinator"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "subjects"
        ]
      }
    },
...
  },
  "info": {
    "title": "School Man API",
    "description": "Generated OpenAPI specification for School Man backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "nationalId": {
            "type": "string",
            "example": "CC-1234567",
            "description": "Unique identifier used nationally for the student"
          },
          "firstName": {
            "type": "string",
            "example": "Juana",
            "maxLength": 80
          },
          "lastName": {
            "type": "string",
            "example": "García",
            "maxLength": 80
          },
          "dob": {
            "type": "string",
            "example": "2010-05-15",
            "description": "ISO-8601 birth date (YYYY-MM-DD)"
          },
          "address": {
            "type": "string",
            "example": "Calle 123 #45-67"
          },
          "guardianName": {
            "type": "string",
            "example": "Marta Ramirez",
            "description": "Full name of the legal guardian"
          },
          "guardianRelationship": {
            "type": "string",
            "example": "Mother",
            "description": "Relationship between guardian and student"
          },
          "guardianPhone": {
            "type": "string",
            "example": "+57 3001234567",
            "description": "Guardian contact phone number"
          }
        },
        "required": [
          "nationalId",
          "firstName",
          "lastName",
          "guardianName",
          "guardianRelationship",
          "guardianPhone"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSubjectAreaDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "MATH",
            "description": "Unique area code",
            "maxLength": 16
          },
          "name": {
            "type": "string",
            "example": "Mathematics",
            "maxLength": 120
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "UpdateSubjectAreaDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSubjectDto": {
        "type": "object",
        "properties": {
          "areaId": {
            "type": "number",
            "example": 1,
            "description": "Identifier of the subject area"
          },
          "code": {
            "type": "string",
            "example": "MATH_ALG",
            "description": "Unique subject code"
          },
          "name": {
            "type": "string",
            "example": "Algebra I",
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "example": "Introduction to algebraic expressions and equations"
          }
        },
        "required": [
          "areaId",
          "code",
          "name"
        ]
      },
      "UpdateSubjectDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}